+++ proto/python/src/soil_power_sensor_protobuf/proto/decode.py
@@ -27,13 +27,14 @@ def decode_response(data: bytes):
     return resp.resp
 
 
-def decode_measurement(data: bytes) -> dict:
+def decode_measurement(data: bytes, raw: bool = True) -> dict:
     """Decodes a Measurement message
 
     The data is decoded into a flat dictionary with the measurement type.
 
     Args:
         data: Byte array of Measurement message.
+        raw: Flag to return raw or adjusted measurements
 
     Returns:
         Flat dictionary of values from the meta field, measurement field, and
@@ -66,11 +67,18 @@ def decode_measurement(data: bytes) -> dict:
     # store measurement data
     meta_dict["data"] = measurement_dict
 
+    # process raw
+    if not raw:
+        # convert measurements to hPa, C, and %
+        if meta_dict["type"] == "bme280":
+            meta_dict["data"]["pressure"] /= 10.0
+            meta_dict["data"]["temperature"] /= 100.0
+            meta_dict["data"]["humidity"] /= 1000.0
+
     # store measurement type
     meta_dict["data_type"] = {}
     for key, value in measurement_dict.items():
         meta_dict["data_type"][key] = type(value)
-
     return meta_dict