+++ proto/python/src/soil_power_sensor_protobuf/proto/encode.py
@@ -139,6 +139,58 @@ def encode_phytos31_measurement(
     # phytos31
     meas.phytos31.voltage = voltage
     meas.phytos31.leaf_wetness = leaf_wetness
+    return meas.SerializeToString()
+
+    return meas.SerializeToString()
+
+
+def encode_bme280_measurement(
+    ts: int,
+    cell_id: int,
+    logger_id: int,
+    pressure: int,
+    temperature: int,
+    humidity: int,
+):
+    """Encodes a BME280Measurement within the Measurement message
+
+    The following raw values correspond to the following SI units
+
+    *Raw*
+
+    pressure: 98473
+    temperature: 2275
+    humidity: 43600
+
+    *SI Units*
+
+    pressure: 9847.3 hPa
+    temperature: 22.75 C
+    humidity: 43.600 %
+
+    Args:
+        ts: Timestamp in unix epochs
+        cell_id: Cell Id from Dirtviz
+        logger_id: Logger Id from Dirtviz
+        pressure: Ambient pressure
+        temperature: Ambient temperature
+        humidity: Relative humidity
+
+    Returns:
+        Serialized BME280 Measurement
+    """
+
+    meas = Measurement()
+
+    # metadata
+    meas.meta.ts = ts
+    meas.meta.cell_id = cell_id
+    meas.meta.logger_id = logger_id
+
+    # bme280
+    meas.bme280.pressure = pressure
+    meas.bme280.temperature = temperature
+    meas.bme280.humidity = humidity
 
     return meas.SerializeToString()